+++ Sasquatch/SasquatchObjC/AppCenterDelegateObjC.mm
@@ -71,7 +71,7 @@
 }
 
 - (void)startAnalyticsFromLibrary {
-  [MSACAppCenter startFromLibraryWithServices:@ [[MSACAnalytics class]]];
+  [MSACAppCenter startFromLibraryWithServices:@[ [MSACAnalytics class] ]];
 }
 
 - (void)setUserId:(NSString *)userId {
@@ -80,6 +80,14 @@
 
 - (void)setCountryCode:(NSString *)countryCode {
   [MSACAppCenter setCountryCode:countryCode];
+}
+
+- (BOOL)isNetworkRequestsAllowed {
+  return [MSACAppCenter isNetworkRequestsAllowed];
+}
+
+- (void)setNetworkRequestsAllowed:(BOOL)isAllowed {
+  [MSACAppCenter setNetworkRequestsAllowed:isAllowed];
 }
 
 #pragma mark - Modules section.
@@ -168,6 +176,18 @@
 
 - (void)generateTestCrash {
   return [MSACCrashes generateTestCrash];
+}
+
+- (void)trackError:(NSError *)error
+    withProperties:(NSDictionary<NSString *, NSString *> *)properties
+       attachments:(NSArray<MSACErrorAttachmentLog *> *)attachments {
+  [MSACCrashes trackError:error withProperties:properties attachments:attachments];
+}
+
+- (void)trackException:(MSACExceptionModel *)exceptionModel
+        withProperties:(NSDictionary<NSString *, NSString *> *)properties
+           attachments:(NSArray<MSACErrorAttachmentLog *> *)attachments {
+  [MSACCrashes trackException:exceptionModel withProperties:properties attachments:attachments];
 }
 
 #pragma mark - MSACDistribute section.