+++ SasquatchMac/SasquatchMacSwift/AppCenterDelegateSwift.swift
@@ -41,6 +41,12 @@ class AppCenterDelegateSwift : AppCenterDelegate {
   func setLogUrl(_ logUrl: String?) {
     AppCenter.logUrl = logUrl;
   }
+  func isNetworkRequestsAllowed() -> Bool {
+    return AppCenter.networkRequestsAllowed
+  }
+  func setNetworkRequestsAllowed(_ isAllowed: Bool) {
+    AppCenter.networkRequestsAllowed = isAllowed
+  }
 
   //MARK: Modules section.
   func isAnalyticsEnabled() -> Bool {
@@ -93,6 +99,14 @@ class AppCenterDelegateSwift : AppCenterDelegate {
   }
   func generateTestCrash() {
     Crashes.generateTestCrash()
+  }
+    
+  func trackError(_ error: Error, withProperties: Dictionary<String, String>?, attachments: [ErrorAttachmentLog]?) {
+    Crashes.trackError(error, properties: withProperties, attachments:attachments)
+  }
+  
+  func trackException(_ exceptionModel: ExceptionModel, withProperties: Dictionary<String, String>?, attachments: [ErrorAttachmentLog]?) -> Void {
+    Crashes.trackException(exceptionModel, properties: withProperties, attachments:attachments)
   }
 
   //MARK: Last crash report section.